requirements.txt
openai
sentence-transformers
faiss-cpu
PyPDF2
scikit-learn
